requirements.txt
#requirements.txt
Flask>=2.2.5
requests>=2.31.0
pytest>=7.3.1
coverage>=7.3.1
python-dotenv>=1.0.0
Flask-WTF>=1.1.1
Werkzeug>=2.2.3
Flask-Limiter>=2.0.1
Flask-Caching>=1.10.1
cachelib>=0.2.1

#pip install -r requirements.txt